requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
Django==5.0.1
django-filter==23.5
djangorestframework==3.14.0
frozenlist==1.4.1
idna==3.6
Markdown==3.5.2
multidict==6.0.4
paypal-payouts-sdk==1.0.0
paypalhttp==1.0.1
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
pytz==2023.4
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
twilio==8.11.1
typing_extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4

# mock==1.0.1;python_version<'3'
# enum34==1.1.6;python_version<'3'
# ipaddress==1.0.19;python_version<'3'
# pyasn1>0.4.2
# vcrpy>1.11.1
# ndg-httpsclient>0.4.4
# m2crypto>0.30.1
